Fix NoMethodError with redmine_contacts (#49)
RedmineTags::TagsHelper#tag_color assumes it will be passed a Tag (or,
to be more specific, something which responds to :name). Now, if the
object does not respond to :name, it simply uses the MD5 hexdigest of
the object itself -- meaning that it will work when passed a String.

(yes, this is a little bit of a hack)